Display on Workout works
The display of the array is working on the workout page. But still need to male the start queue button work and the display needs to show timer going done. the sequential nature of display also needs to be added. Once this is done then will add the remaining timers too.