The Countdown is Depreceating
The countdown timer is deprecating by 2 on the display. I think its because of the lag between working and displaying. I tried to debug it but its not working. I am basically updating the array time value and I think This is a mistake. I am planning to sit with the professor to understand the next steps as I feel I have reached a road block.